+++ packages/app-andrew/src/pages/group/[slug].tsx
@@ -18,6 +18,7 @@ import { FavsButton } from 'components/group/FavsButton'
 import { SettingsButton } from 'components/group/SettingsButton'
 import { JoinLeaveButton } from 'components/group/JoinLeaveButton'
 import { GroupMembers } from 'components/group/GroupMembers'
+import { FollowButton } from 'components/group/FollowButton'
 // ui
 import { ErrorInfo } from '@gtms/ui/ErrorInfo'
 import { NavigationTabs } from '@gtms/ui/NavigationTabs'
@@ -145,6 +146,7 @@ const GroupPage: NextPage<GroupPageProps> = (props) => {
                 <FavsButton group={state.group} />
                 <JoinLeaveButton group={state.group} />
                 <SettingsButton group={state.group} />
+                <FollowButton group={state.group} />
               </div>
             </div>
           </div>