+++ packages/state-group/src/groupsList/index.ts
@@ -1,0 +1,3 @@
+export * from './groupsList.query'
+export * from './groupsList.service'
+export * from './groupsList.store'