+++ packages/api-notifications/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/api-notifications@0.1.0...@gtms/api-notifications@0.2.0) (2020-06-21)
+
+
+### Features
+
+* GEOT-287: follow a group button ([#660](https://github.com/gtms-org/gtms-frontend/issues/660)) ([b30c5e6](https://github.com/gtms-org/gtms-frontend/commit/b30c5e6b82013f35c29f4f9335bcd50c92fcb8a3))
+
+
+
+
+
 # 0.1.0 (2020-06-21)